+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add a String type only JSON Schema class
 - Add a Boolean type only JSON Schema class
+- Add a Number type only JSON Schema class
 
 
 ## 1.0.0-dev.9